+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# 3.9.1
+## Feature
+## Bug Fix
+  - Bug: SQL ไม่ได้ระบุ first() ในการนับจำนวน
+
+# 3.9.0
+## Feature
+  - เพิ่ม Feature สำหรับการประเมินความพึงพอใจใน MOPH Alert
+  - เพิ่มการส่งข้อมูลการรับบริการ getVisitForMophAlert เพื่อใช้ประเมินความพึงพอใจ
+  - เพิ่ม SQL Lite เพื่อบันทึกข้อมูลการส่ง MOPH Alert
+  - แก้ไขการส่ง MOPH Alert เป็นที่ละ 100 rows แทนตัวเดิมที่ส่งทีละ 500 rows
+
+# 3.8.14
+## Feature
+  - เพิ่ม source HIS: VPM
+  - เปลี่ยนตัวแปร Key ให้สอดคล้องกับ MOPH ERP
+  - เพิ่ม getVisitForMophAlert
+  - เพิ่มเงื่อนไขใน HOSxP ให้ sum เฉพาะ ward.ward_active=Y
+  - ยกเลิกการส่ง spclty.name ใน HOSxP
+  - ยกเลิกการประมวลผล OPD Visit ย้อนหลัง
+  - ปรับ SQL HOSxP จากตาราง ipt_spclty เป็น spclty
+
 # 3.8.13
 ## Feature
   - แก้ไข SQL สำหรับ HospitalOS กรณี return 204 และ 413